Moved ingredient validation functions into ingredient.ts because I want to try grouping by recipe components (yield, ingredient, etc.) instead of by functionality (form editing, viewing). I think this makes more sense. I'm not really sure
