+++ src/core/form-validation.ts
@@ -1,8 +1,5 @@
 import type { RecipeBook } from "./recipe";
-
-export const isNumeric = (amount: string): boolean => {
-  return amount !== "" && !isNaN(Number(amount));
-};
+import { isNumeric } from "./utils";
 
 /**
  * Evaluate whether newName is a valid recipe name. If it is not,
@@ -59,40 +56,6 @@ export const validateRecipeYieldUnits = (units: string): string => {
     return "Recipe yield units cannot be empty";
   } else if (units.length > 25) {
     return "Recipe yield units cannot be longer than 25 characters";
-  } else {
-    return "";
-  }
-};
-
-/**
- * Checks whether the ingredient name is a valid string. If it isn't, returns
- * an error message describing the error. If it is, returns an empty string.
- * @param name - The proposed ingredient name
- * @returns The error message, or an empty string if there are no errors
- */
-export const validateIngredientName = (name: string): string => {
-  if (name.length === 0) {
-    return "Ingredient name cannot be empty";
-  } else if (name.length > 50) {
-    return "Ingredient name cannot be longer than 50 characters";
-  } else {
-    return "";
-  }
-};
-
-/**
- * Checks whether the ingredient amount is a valid number. If it isn't, returns
- * an error message describing the error. It if is, returns an empty string.
- * @param amount - The proposed ingredient amount
- * @returns The error message, or an empty string if there are no errors
- */
-export const validateIngredientAmount = (amount: string): string => {
-  if (amount === "") {
-    return "Ingredient amount is required";
-  } else if (!isNumeric(amount)) {
-    return "Ingredient amount must be a number";
-  } else if (Number(amount) <= 0) {
-    return "Ingredient amount must be greater than zero";
   } else {
     return "";
   }
