+++ src/core/utils.ts
@@ -1,0 +1,3 @@
+export const isNumeric = (amount: string): boolean => {
+  return amount !== "" && !isNaN(Number(amount));
+};
